# Simplify stack