--- a/src/ConnectionPool.jl
+++ b/src/ConnectionPool.jl
@@ -119,7 +119,7 @@ same parameters, we can re-use it (as long as it's not already being used, obvio
 function hashconn end
 
 hashconn(x::Connection{T}) where {T} = hashconn(T, x.host, x.port, x.pipeline_limit, x.require_ssl_verification, x.clientconnection)
-hashconn(T, host, port, pipeline_limit, require_ssl_verification, client) = hash(T, hash(host, hash(port, hash(pipeline_limit, hash(require_ssl_verification, hash(client, UInt(0)))))))
+hashconn(T::Type, host, port, pipeline_limit, require_ssl_verification, client) = hash(T, hash(host, hash(port, hash(pipeline_limit, hash(require_ssl_verification, hash(client, UInt(0)))))))
 
 """
     Transaction