--- a/src/TopRequest.jl
+++ b/src/TopRequest.jl
@@ -12,7 +12,7 @@ It's useful for inserting a custom layer at the top of the stack.
 """
 abstract type TopLayer{Next <: Layer} <: Layer{Next} end
 
-request(::Type{TopLayer{Next}}, args...; kwargs...) where Next =
+request(::Type{TopLayer{Next}}, args...; kwargs...) =
     request(Next, args...; kwargs...)
 
 end