@c42f 2021-08-05
So the main change here is that the stack is no longer a big nested layer type, but a linked list of `Stack{LayerType}` structs, with `next` pointer of type `Stack` which will be dispatched on at runtime.

This is a pretty large change to the original design which very explicitly aimed to avoid runtime dispatch. Maybe that's fine in the name of simplifying things and I've wondered about this myself.

But I think testing against `example.com` is not going to tell you anything useful — mostly you're measuring network latency and perhaps also performance/load on the remote server.  Rather, it's better to test against a high performance server running on localhost.

Unfortunately we don't have the original author of the current design around anymore to ask for advice and motivating examples.