@rikhuijzer 2021-08-05
Thanks for taking a look at the PR. I just figured that my whole PR is based on a wrong assumption. I thought that the `Next` part of the methods was always dispatching on the full stacked type and therefore causing many unnecessary methodinstances. That's not true, so performance wise, the implementation was already very good. 

I could still rewrite this PR to simplify the huge ternary statement in `stack` and and the TopLayer, but I probably won't. So, I'm closing this for now.