@c42f 2021-08-06
> I thought that the Next part of the methods was always dispatching on the full stacked type and therefore causing many unnecessary methodinstances.

Ah I see.  Well, this isn't entirely true, but it may be partially true — while *dispatch* happens based on the individual layer types, *specialization* may happen based on the concrete type of the stack from the current layer downward (though the exact amount of specialization is an implementation detail of the compiler).

Personally I think it would be interesting and useful to change the stack into a stack of layer *instances* rather than types. Then the layers can contain state if necessary (for example, the connection pool layer would no longer need to rely on global state).

If you do want to consider this design further, I'd suggest abstracting the stack such that implementation of individual layers is decoupled from the concrete representation of the stack. Then users could choose their stack representation to hide information from the compiler (as in your `Stack` list) or give it all the information as in the current highly nested stack type.