@rikhuijzer 2021-08-06
> [...] _specialization_ may happen based on the concrete type of the stack from the current layer downward (though the exact amount of specialization is an implementation detail of the compiler).

Fair enough. Well, in Julia 1.7 it looked fine when inspecting `methods(requests)` via `MethodAnalysis.jl`.

> If you do want to consider this design further,

I don't think anymore that there is much to gain here, so probably not.